interface

+++ SAE/SAE_Exoplanete/src/clustering/Clustering.java
@@ -1,5 +1,5 @@
 package clustering;
 
 public interface Clustering {
-    int[] clusteriser(int[][] objets);
+    int[] clusteriser(double[][] objets);
 }
